Settings to own file
